chore: add common file for all property wrappers
